优化及修复游戏配置模块

+++ src/main/java/com/zhlzzz/together/controllers/game/GameConfigController.java
@@ -1,5 +1,6 @@
 package com.zhlzzz.together.controllers.game;
 
+import com.zhlzzz.together.controllers.ApiAuthentication;
 import com.zhlzzz.together.controllers.ApiExceptions;
 import com.zhlzzz.together.game.GameConfigImpl;
 import com.zhlzzz.together.game.GameType;
@@ -8,6 +9,8 @@ import com.zhlzzz.together.game.game_config.GameConfig;
 import com.zhlzzz.together.game.game_config.GameConfigEntity;
 import com.zhlzzz.together.game.game_config.GameConfigOptionEntity;
 import com.zhlzzz.together.game.game_config.GameConfigService;
+import com.zhlzzz.together.user.User;
+import com.zhlzzz.together.user.UserService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
@@ -27,16 +30,23 @@ public class GameConfigController {
 
     private final GameConfigService gameConfigService;
     private final GameTypeService gameTypeService;
+    private final UserService userService;
 
     @PostMapping
     @ApiOperation(value = "新增游戏配置")
     @ResponseBody
     public GameConfigView createGameConfig(@RequestParam Integer gameTypeId, @RequestParam GameConfig.InputType inputType,
                                            @RequestParam String label, @RequestParam(required = false, defaultValue = "true")
-                                                       boolean required) {
+                                                       boolean required, ApiAuthentication auth) {
+
+        User admin = userService.getUserById(auth.requireUserId()).filter(u -> u.isAdmin()).orElse(null);
+        if (admin == null) {
+            throw ApiExceptions.noPrivilege();
+        }
+
         GameType gameType = gameTypeService.getGameTypeById(gameTypeId).orElseThrow(()-> ApiExceptions.notFound("不存在此游戏类型"));
 
-        GameConfigEntity gameConfig = gameConfigService.addGameConfig(gameTypeId, inputType, label, required);
+        GameConfigEntity gameConfig = gameConfigService.addGameConfig(gameType.getId(), inputType, label, required);
 
         return new GameConfigView(new GameConfigImpl(gameConfig, new ArrayList<>()));
     }
@@ -46,7 +56,11 @@ public class GameConfigController {
     @ResponseBody
     public GameConfigView updateGameConfig(@PathVariable Long id, @RequestParam GameConfig.InputType inputType,
                                            @RequestParam String label, @RequestParam(required = false, defaultValue = "true")
-                                                       boolean required) {
+                                                       boolean required, ApiAuthentication auth) {
+        User admin = userService.getUserById(auth.requireUserId()).filter(u -> u.isAdmin()).orElse(null);
+        if (admin == null) {
+            throw ApiExceptions.noPrivilege();
+        }
         GameConfigEntity gameConfig = gameConfigService.updateGameConfig(id,inputType,label,required);
 
         return new GameConfigView(new GameConfigImpl(gameConfig, new ArrayList<>()));
@@ -55,14 +69,22 @@ public class GameConfigController {
     @DeleteMapping(path = "/{id:\\d+}")
     @ApiOperation(value = "删除游戏配置")
     @ResponseBody
-    public void deleteGameConfig(@PathVariable Long id) {
+    public void deleteGameConfig(@PathVariable Long id, ApiAuthentication auth) {
+        User admin = userService.getUserById(auth.requireUserId()).filter(u -> u.isAdmin()).orElse(null);
+        if (admin == null) {
+            throw ApiExceptions.noPrivilege();
+        }
         gameConfigService.deleteGameConfig(id);
     }
 
-    @PostMapping(path = "/id:\\d+/options")
+    @PostMapping(path = "/{id:\\d+}/options")
     @ApiOperation(value = "新增游戏配置选项")
     @ResponseBody
-    public GameConfigView.Option createOption(@PathVariable Long id, @RequestParam String value) {
+    public GameConfigView.Option createOption(@PathVariable Long id, @RequestParam String value, ApiAuthentication auth) {
+        User admin = userService.getUserById(auth.requireUserId()).filter(u -> u.isAdmin()).orElse(null);
+        if (admin == null) {
+            throw ApiExceptions.noPrivilege();
+        }
         GameConfigEntity gameConfig= gameConfigService.getGameConfigById(id).orElseThrow(()-> ApiExceptions.notFound("不存在此游戏配置"));
 
         GameConfig.Option option = gameConfigService.addOption(gameConfig.getId(), value);
@@ -70,20 +92,28 @@ public class GameConfigController {
         return new GameConfigView.Option(option);
     }
 
-    @PutMapping(path = "/configId:\\d+/options/id:\\d+")
-    @ApiOperation(value = "新增游戏配置选项")
+    @PutMapping(path = "/{configId:\\d+}/options/{id:\\d+}")
+    @ApiOperation(value = "更新游戏配置选项")
     @ResponseBody
-    public GameConfigView.Option updateOption(@PathVariable Long id, @RequestParam String value) {
+    public GameConfigView.Option updateOption(@PathVariable Long id, @RequestParam String value, ApiAuthentication auth) {
+        User admin = userService.getUserById(auth.requireUserId()).filter(u -> u.isAdmin()).orElse(null);
+        if (admin == null) {
+            throw ApiExceptions.noPrivilege();
+        }
         GameConfigOptionEntity gameConfigOption = gameConfigService.getGameConfigOptionById(id).orElseThrow(()-> ApiExceptions.notFound("不存在此游戏配置选项"));
         GameConfig.Option option = gameConfigService.updateOption(gameConfigOption.getId(), value);
 
         return new GameConfigView.Option(option);
     }
 
-    @DeleteMapping(path = "/configId:\\d+/options/id:\\d+")
+    @DeleteMapping(path = "/{configId:\\d+}/options/{id:\\d+}")
     @ApiOperation(value = "删除游戏配置选项")
     @ResponseBody
-    public void deleteGameConfigOption(@PathVariable Long id) {
+    public void deleteGameConfigOption(@PathVariable Long id, ApiAuthentication auth) {
+        User admin = userService.getUserById(auth.requireUserId()).filter(u -> u.isAdmin()).orElse(null);
+        if (admin == null) {
+            throw ApiExceptions.noPrivilege();
+        }
         gameConfigService.deleteOption(id);
     }
